Add files via upload
I have used the web socket (Rachet) to link the data from the temperature sensor and send it to PHP. In the following code, I utilized the WebSocketsClient library to establish a connection between an ESP32 device and a server over Wi-Fi. The temperature sensor data, read from an analog input pin, is continuously sent to the server via a WebSocket connection every 30 seconds. The code begins by connecting to a Wi-Fi network and then establishes a WebSocket connection to the server using the IP address and port. Once connected, it sends the temperature data in the form of a message string to the server.